+++ src/main/java/server/poptato/external/firebase/application/FcmNotificationService.java
@@ -1,4 +1,4 @@
-package server.poptato.external.firebase.service;
+package server.poptato.external.firebase.application;
 
 import com.google.firebase.messaging.FirebaseMessagingException;
 import com.google.firebase.messaging.MessagingErrorCode;
